[write] Reorganize gpos/gsub modules
Specifically adds gpos/ and gsub/ directories and moves some stuff into
those that we were previously doing hacky #[path] crap with.

This is no functional change, but I will be moving the gpos/gsub
builders from fea-rs into these directories subsequently.
